Changed SLS services API

+++ Phylotastic_SLS_WebServices/Phylotastic_SLS_WebServices_Running.py
@@ -14,7 +14,9 @@ import collections
 from cherrypy import tools
 
 from support import species_list_service
-#from support import extract_names_service
+from support import authenticate_user
+from distutils.util import strtobool
+import types
 
 from __builtin__ import True
 
@@ -30,26 +32,15 @@ PUBLIC_HOST_ROOT_WS = "http://%s/%s" %(str(IP_ADDRESS),str(WS_NAME))
 ACCESS_LOG_CHERRYPY_5005 = ROOT_FOLDER + "/log/%s_5005_access_log.log" %(str(datetime.datetime.fromtimestamp(time.time()).strftime('%Y%m%d')))
 ERROR_LOG_CHERRYPY_5005 = ROOT_FOLDER + "/log/%s_5005_error_log.log" %(str(datetime.datetime.fromtimestamp(time.time()).strftime('%Y%m%d')))
 
-
-#Prefix Filename
-
-class MultipleLevelsOfDictionary(collections.OrderedDict):
-    def __getitem__(self,item):
-        try:
-            return collections.OrderedDict.__getitem__(self,item)
-        except:
-            value = self[item] = type(self)()
-            return value
-
-def return_response_error(code,type,mess,format="JSON"):
-    if (format=="JSON"):
-        cherrypy.response.headers['Content-Type'] = "application/json"
-        cherrypy.response.headers['Retry-After']=60
-        cherrypy.response.status = code
-        message = {type:mess}
-        return json.dumps(message)
+def return_response_error(error_code, error_message,response_format="JSON"):
+    error_response = {'message': error_message, 'status_code':error_code}
+    if (response_format == "JSON"):
+        #cherrypy.response.headers['Content-Type'] = "application/json"
+        #cherrypy.response.headers['Retry-After']=60
+        #cherrypy.response.status = error_code    
+        return json.dumps(error_response)
     else:
-        return "Not support yet"
+        return error_response
     
 def is_json(myjson):
     try:
@@ -65,71 +56,37 @@ class Species_List_Service_API(object):
         return "Species_List_Service API (Abu Saleh) : Gives access to user's lists of species";
     #---------------------------------------------
     #---------------------------------------------
-    def get_user_lists(self, include_all=False, **request_data):
-        try:
- 			user_id = str(request_data['user_id']).strip();
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
-        conn = species_list_service.connect_mongodb()
-        service_result = species_list_service.get_user_lists(user_id, conn, include_all)   
-        
-        return service_result;
+    def get_list(self, verbose=False, content=True, list_id="-1", user_id=None, access_token=None):
+        if type(content) != types.BooleanType:
+            content = strtobool(content)
+        if type(verbose) != types.BooleanType:
+            verbose = strtobool(verbose)
 
- 	#------------------------------------------------
-    def get_public_lists(self, include_all=False):
         conn = species_list_service.connect_mongodb()
-        service_result = species_list_service.get_public_lists(conn, include_all)   
-        
-        return service_result;
+        service_result = species_list_service.get_list(conn, user_id, int(list_id), verbose, content, access_token)   
+        conn.close()
 
-    #------------------------------------------------
-    def get_list(self, include_all=False, **request_data):
-        try:
- 			list_id = str(request_data['list_id']).strip();
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
-        conn = species_list_service.connect_mongodb()
-        service_result = species_list_service.get_list(int(list_id), conn, include_all)   
-        
         return service_result;
-
-    #------------------------------------------------
-    def get_species(self, include_all=False, **request_data):
-        try:
- 			list_id = str(request_data['list_id']).strip();
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
-        conn = species_list_service.connect_mongodb()
-        service_result = species_list_service.get_list_species(int(list_id), conn, include_all)   
-        
-        return service_result;
-
+    
     #-----------------------------------------------------
-    def find_lists(self, include_all=False, user_id=-1, **request_data):
-        try:
- 			search_query = str(request_data['keyword']).strip();
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
-        conn = species_list_service.connect_mongodb()
-        service_result = species_list_service.find_lists(conn, search_query, user_id, include_all)   
-        
-        return service_result;
-
- 	#--------------------------------------------
     def remove_list(self, **request_data):
         try:
- 			user_id = str(request_data['user_id']).strip();
- 			list_id = str(request_data['list_id']).strip();
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
+ 			user_id = str(request_data['user_id']).strip()
+ 			list_id = str(request_data['list_id']).strip()
+ 			access_token = str(request_data['access_token']).strip()
+
+        except Exception, e:
+            return return_response_error(400,"Error: Missing parameter %s"%(str(e)),"JSON")
+
+        #verify user
+        token_verification = authenticate_user.verify_access_token(access_token, user_id)
+        if not(token_verification['is_access_token_valid']):
+           return return_response_error(400,"Error:"+ token_verification['message'],"JSON")
+        #allow access to service for verified user
         conn = species_list_service.connect_mongodb()
         service_result = species_list_service.remove_user_list(user_id, int(list_id), conn)   
-        
+        conn.close()
+
         return service_result;
 
   	#---------------------------------------------
@@ -138,14 +95,22 @@ class Species_List_Service_API(object):
     def replace_species(self,**request_data):
         try:
             input_json = cherrypy.request.json
-            #if not (is_json(input_json)):
- 			#	return return_response_error(400,"error","content-type must be application/json","JSON") 				 
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
+            access_token = input_json["access_token"]
+            user_id =  input_json["user_id"]
+        except KeyError, e:
+            return return_response_error(400,"Error: Missing parameter %s"%(str(e)),"NotJSON")     		 
+        except Exception, e:
+            return return_response_error(400,"Error:" + str(e), "NotJSON")
+
+        #verify user
+        token_verification = authenticate_user.verify_access_token(access_token, user_id)
+        if not(token_verification['is_access_token_valid']):
+            return return_response_error(400,"Error:"+ token_verification['message'],"NotJSON")
+        #allow access to service for verified user
         conn = species_list_service.connect_mongodb()
         service_result = species_list_service.replace_species_in_list(input_json, conn)   
-           
+        conn.close()
+   
         return service_result;
 
   	#-----------------------------------------------	
@@ -154,29 +119,39 @@ class Species_List_Service_API(object):
     def remove_species(self,**request_data):
         try:
             input_json = cherrypy.request.json
-             
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
+            access_token = input_json["access_token"]
+            user_id =  input_json["user_id"]
+        except KeyError, e:
+            return return_response_error(400,"Error: Missing parameter %s"%(str(e)),"NotJSON")     		 
+        except Exception, e:
+            return return_response_error(400,"Error:" + str(e), "NotJSON")
+
+        #verify user
+        token_verification = authenticate_user.verify_access_token(access_token, user_id)
+        if not(token_verification['is_access_token_valid']):
+            return return_response_error(400,"Error:"+ token_verification['message'],"NotJSON")
+        #allow access to service for verified user
         conn = species_list_service.connect_mongodb()
         service_result = species_list_service.remove_species_from_list(input_json, conn)   
-           
-        return service_result;
- 	#------------------------------------------------
-    
+        conn.close()
+   
+        return service_result; 	
+ 	#------------------------------------------------   
     @cherrypy.tools.json_out()
     @cherrypy.tools.json_in()
     def insert_list(self,**request_data):
         try:
             input_json = cherrypy.request.json
-            	 
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
-        
+            #if not (is_json(input_json)):
+ 			#	return return_response_error(400,"error","content-type must be application/json","JSON") 				 
+        except Exception, e:
+            return return_response_error(400,"Error:" + str(e), "NotJSON")
+
         conn = species_list_service.connect_mongodb()
         service_result = species_list_service.insert_user_list(input_json, conn)   
-           
-        return service_result;
+        conn.close()
+   
+        return service_result
  	#-----------------------------------------------
 
     @cherrypy.tools.json_out()
@@ -186,27 +161,23 @@ class Species_List_Service_API(object):
             input_json = cherrypy.request.json
             #if not (is_json(input_json)):
  			#	return return_response_error(400,"error","content-type must be application/json","JSON") 				 
-        except:
-            return return_response_error(400,"error","Missing parameters text","JSON")
+        except Exception, e:
+            return return_response_error(400,"Error:" + str(e), "NotJSON")
         
         conn = species_list_service.connect_mongodb()
         service_result = species_list_service.insert_species_to_list(input_json, conn)   
-           
+        conn.close()   
         return service_result;
 
 
     #Public /index
     index.exposed = True
-    get_user_lists.exposed = True
-    get_public_lists.exposed = True
     get_list.exposed = True
-    find_lists.exposed = True
-    get_species.exposed = True
     insert_list.exposed = True
     remove_list.exposed = True
     insert_species.exposed = True
-    replace_species.exposed = True
     remove_species.exposed = True
+    replace_species.exposed = True
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 #--------------------------------------------------------------------------------------------
